修改删除文件时，文件路径为空而导致程序崩溃

+++ android/common/src/main/java/com/example/file/FileUtils.java
@@ -271,7 +271,14 @@ public class FileUtils {
 
 
     public static boolean deleteFileByPath(String filePath) {
+
+        if (filePath == null) {
+            return false;
+        }
+
         File deleteFile = new File(filePath);
+
+
         if (deleteFile.exists()) {
             return deleteFile.delete();
         }
